refactor(style-config): rename tailwindconfig to style-config & add global css
