+++ apps/web/tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 
-const base = require("tailwindconfig")
+const base = require("style-config")
 
 module.exports = {
   content: [
