+++ packages/style-config/README.md
@@ -1,0 +1,1 @@
+# style-config
