+++ packages/style-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tailwindconfig",
+  "name": "style-config",
   "main": "index.js",
   "version": "0.0.0",
   "devDependencies": {
