+++ packages/style-config/semantic/infos.js
