+++ packages/ui/cosmos.config.json
@@ -1,4 +1,3 @@
 {
-  "plugins": ["react-cosmos-plugin-vite"],
-  "globalImports": ["./reset.css"]
+  "plugins": ["react-cosmos-plugin-vite"]
 }
