+++ packages/ui/package.json
@@ -31,7 +31,7 @@
     "react-cosmos": "next",
     "react-cosmos-plugin-vite": "next",
     "react-dom": "latest",
-    "tailwindconfig": "*",
+    "style-config": "*",
     "tailwindcss": "^3.2.6",
     "tsconfig": "*",
     "typescript": "^5.3.2",
