+++ packages/ui/src/TagsList/TagsList.module.css
@@ -1,6 +1,6 @@
 .list {
   @apply flex flex-row justify-start items-center gap-x-2;
-  @apply w-full overflow-auto no-scrollbar;
+  @apply w-full overflow-auto no-scrollbar py-1;
 }
 
 .tag {
