+++ packages/ui/tailwind.config.cjs
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 
-const base = require('tailwindconfig')
+const base = require('style-config')
 
 module.exports = {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
